# src/BookBarter.React/src/api/generated/models/get-paged-publishers-for-moderation-query.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BookBarter.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface GetPagedPublishersForModerationQuery
 */
export interface GetPagedPublishersForModerationQuery {
    /**
     * 
     * @type {number}
     * @memberof GetPagedPublishersForModerationQuery
     */
    'pageSize'?: number;
    /**
     * 
     * @type {number}
     * @memberof GetPagedPublishersForModerationQuery
     */
    'pageNumber'?: number;
    /**
     * 
     * @type {string}
     * @memberof GetPagedPublishersForModerationQuery
     */
    'orderByProperty'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof GetPagedPublishersForModerationQuery
     */
    'orderDirection'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof GetPagedPublishersForModerationQuery
     */
    'approved'?: boolean;
}